+++ helper/decoder/influxdb/decoder.go
@@ -88,14 +88,16 @@ func (d *Decoder) parsePointsToLogs(points []models.Point, req *http.Request) []
 		if err != nil {
 			continue
 		}
-		var valueType = valueTypeFloat
+		var valueType string
 		var value string
 		for field, v := range fields {
 			switch v := v.(type) {
 			case float64:
 				value = strconv.FormatFloat(v, 'g', -1, 64)
+				valueType = valueTypeFloat
 			case int64:
 				value = strconv.FormatInt(v, 10)
+				valueType = valueTypeInt
 			case bool:
 				if v {
 					value = "1"