+++ pkg/protocol/converter/converter_metric_test.go
@@ -176,13 +176,51 @@ func Test_metricReader_readSortedLabels(t *testing.T) {
 			reader: &metricReader{
 				labels: "bb#$#aa|aa#$#bb|",
 			},
-			wantErr: true,
+			wantLabels: []metricLabel{
+				{key: "aa", value: "bb"},
+				{key: "bb", value: "aa"},
+			},
 		},
 		{
 			reader: &metricReader{
 				labels: "bb",
 			},
 			wantErr: true,
+		},
+		{
+			reader: &metricReader{
+				labels: "bb#$#aa|eee",
+			},
+			wantLabels: []metricLabel{
+				{key: "bb", value: "aa|eee"},
+			},
+		},
+		{
+			reader: &metricReader{
+				labels: "bb#$#aa|eee|aa#$#bb",
+			},
+			wantLabels: []metricLabel{
+				{key: "aa", value: "bb"},
+				{key: "bb", value: "aa|eee"},
+			},
+		},
+		{
+			reader: &metricReader{
+				labels: "bb#$#aa||eee||aa#$#bb",
+			},
+			wantLabels: []metricLabel{
+				{key: "aa", value: "bb"},
+				{key: "bb", value: "aa||eee|"},
+			},
+		},
+		{
+			reader: &metricReader{
+				labels: "cc||bb#$#aa||eee||aa#$#bb",
+			},
+			wantLabels: []metricLabel{
+				{key: "aa", value: "bb"},
+				{key: "cc||bb", value: "aa||eee|"},
+			},
 		},
 	}